backend/requirements.txt
fastapi
uvicorn
pydantic
geojsplit
SQLAlchemy
alembic
GeoAlchemy2
geojson-pydantic
PyJWT
dotenv